+++ Cartera/Modelo/MFinanciacion.cs
@@ -23,7 +23,7 @@ namespace Cartera.Modelo
         public int Valor_Interes { get; set; }
         public DateTime Fecha_Recaudo { get; set; }
                                             
-        public static int crearFinanciacion(int Valor_Producto_Financiacion, int Valor_Entrada, int Valor_Sin_interes, int Valor_Cuota_Sin_interes, int Cuotas_Sin_interes, int Valor_Con_Interes, int Cuotas_Con_Interes, int Valor_Cuota_Con_Interes, int Valor_Interes, string Fecha_Recaudo, string Fk_Producto)
+        public static int crearFinanciacion(int Valor_Producto_Financiacion, int Valor_Entrada, int Valor_Sin_interes, double Valor_Cuota_Sin_interes, int Cuotas_Sin_interes, double Valor_Con_Interes, int Cuotas_Con_Interes, double Valor_Cuota_Con_Interes, int Valor_Interes, string Fecha_Recaudo, string Fk_Producto)
         {
             string sql = "INSERT INTO Financiacion (Valor_Producto_Financiacion, Valor_Entrada, Valor_Sin_interes, Valor_Cuota_Sin_interes, Cuotas_Sin_interes, Valor_Con_Interes, Cuotas_Con_Interes, Valor_Cuota_Con_Interes, Valor_Interes, Fecha_Recaudo, Fk_Producto, Estado_Financiacion) VALUES(@Valor_Producto_Financiacion, @Valor_Entrada, @Valor_Sin_interes, @Valor_Cuota_Sin_interes, @Cuotas_Sin_interes, @Valor_Con_Interes, @Cuotas_Con_Interes, @Valor_Cuota_Con_Interes, @Valor_Interes, @Fecha_Recaudo, @Fk_Producto, @Estado_Financiacion);";
             SQLiteCommand cmd = new SQLiteCommand(sql, Conexion.instanciaDb());
@@ -54,7 +54,7 @@ namespace Cartera.Modelo
         {
             return Conexion.consulta("SELECT  Id_Financiacion, Valor_Producto_Financiacion, Valor_Sin_interes , Valor_Entrada, Valor_Cuota_Sin_interes, Cuotas_Sin_interes, Valor_Con_Interes, Cuotas_Con_Interes, Valor_Cuota_Con_Interes, Valor_Interes, Fecha_Recaudo, Id_Refinanciacion FROM Financiacion LEFT JOIN Refinanciacion on Fk_Financiacion=Id_Financiacion WHERE Fk_Producto= '" + id_Producto + "' and Estado_Financiacion='Activa';");
         }
-        public static int actualizarFinanciacion(int Id_Financiacion, int Valor_Producto_Financiacion, int Valor_Entrada, int Valor_Sin_interes, int Valor_Cuota_Sin_interes, int Cuotas_Sin_interes, int Valor_Con_Interes, int Cuotas_Con_Interes, int Valor_Cuota_Con_Interes, int Valor_Interes, string Fecha_Recaudo, int Fk_Producto)
+        public static int actualizarFinanciacion(int Id_Financiacion, int Valor_Producto_Financiacion, int Valor_Entrada, int Valor_Sin_interes, double Valor_Cuota_Sin_interes, int Cuotas_Sin_interes, double Valor_Con_Interes, int Cuotas_Con_Interes, double Valor_Cuota_Con_Interes, int Valor_Interes, string Fecha_Recaudo, int Fk_Producto)
         {
                                               
             string sql = "UPDATE Financiacion SET Valor_Producto_Financiacion = @Valor_Producto_Financiacion, Valor_Entrada = @Valor_Entrada, Valor_Sin_interes = @Valor_Sin_interes, Valor_Cuota_Sin_interes = @Valor_Cuota_Sin_interes, Cuotas_Sin_interes = @Cuotas_Sin_interes, Valor_Con_Interes = @Valor_Con_Interes, Cuotas_Con_Interes = @Cuotas_Con_Interes, Valor_Cuota_Con_Interes = @Valor_Cuota_Con_Interes, Valor_Interes = @Valor_Interes, Fecha_Recaudo = @Fecha_Recaudo, Fk_Producto = @Fk_Producto WHERE Id_Financiacion = '" + Id_Financiacion + "';";
